+++ engine/spec/components/citizens_advice_components/text_input_spec.rb
@@ -22,8 +22,8 @@ RSpec.describe CitizensAdviceComponents::TextInput, type: :component do
       expect(page).to have_selector "#example-input-input"
     end
 
-    it "includes required attribute" do
-      expect(page).to have_selector "input[required]"
+    it "includes aria-required attribute" do
+      expect(page).to have_selector "input[aria-required=true]"
     end
 
     it "does not have an aria-describedby attribute by default" do